// next-auth.d.ts
import { UserRole } from "@prisma/client";
import NextAuth, { type DefaultSession } from "next-auth";

export type ExtendedUser = {
    role: UserRole;
    isTwoFactorEnabled: boolean;
    isOAuth: boolean;
} & DefaultSession["user"]

// declare module "next-auth" {
//     interface Session {
//         user: ExtendedUser;
//     }
// }